feat: enable highlighting of categories
Categories which are marked as `is_highlight` are visually brought out.
In the software edit view they get an extra dropdown menu and in the
software view they have an extra section in the main content area.

Categories on different levels in the hierarchy could have different
extra properties. These are now aggregated in one database column
`properties` of type JSONB.